Remove unnecessary Annotation and related Processor
