+++ shardingsphere-test/shardingsphere-integration-test/shardingsphere-integration-test-fixture/src/test/java/org/apache/shardingsphere/test/integration/junit/container/adapter/ShardingSphereAdapterContainer.java
@@ -17,12 +17,15 @@
 
 package org.apache.shardingsphere.test.integration.junit.container.adapter;
 
+import com.google.common.io.ByteStreams;
 import lombok.Getter;
-import org.apache.shardingsphere.test.integration.junit.annotation.XmlResource;
+import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
+import org.apache.shardingsphere.infra.metadata.user.yaml.config.YamlUserConfiguration;
+import org.apache.shardingsphere.infra.yaml.engine.YamlEngine;
+import org.apache.shardingsphere.proxy.config.yaml.YamlProxyServerConfiguration;
 import org.apache.shardingsphere.test.integration.junit.container.ShardingSphereContainer;
 import org.apache.shardingsphere.test.integration.junit.param.model.ParameterizedArray;
-import org.apache.shardingsphere.test.integration.junit.processor.AuthenticationProcessor;
-import org.apache.shardingsphere.test.integration.junit.processor.AuthenticationProcessor.Authentication;
 
 import javax.sql.DataSource;
 
@@ -32,11 +35,21 @@ import javax.sql.DataSource;
 public abstract class ShardingSphereAdapterContainer extends ShardingSphereContainer {
     
     @Getter
-    @XmlResource(file = "/docker/{scenario}/proxy/conf/server.yaml", processor = AuthenticationProcessor.class)
     private Authentication authentication;
     
     public ShardingSphereAdapterContainer(final String dockerName, final String dockerImageName, final ParameterizedArray parameterizedArray) {
         this(dockerName, dockerImageName, false, parameterizedArray);
+        this.authentication = loadAuthentication(parameterizedArray);
+    }
+    
+    @SneakyThrows
+    private Authentication loadAuthentication(final ParameterizedArray parameterizedArray) {
+        YamlProxyServerConfiguration proxyServerConfiguration = YamlEngine.unmarshal(
+                ByteStreams.toByteArray(this.getClass().getResourceAsStream("/docker/" + parameterizedArray.getScenario() + "/proxy/conf/server.yaml")),
+                YamlProxyServerConfiguration.class
+        );
+        YamlUserConfiguration configuration = proxyServerConfiguration.getAuthentication().getUsers().get("root");
+        return new Authentication("root", configuration.getPassword());
     }
     
     public ShardingSphereAdapterContainer(final String dockerName, final String dockerImageName, final boolean isFakeContainer, final ParameterizedArray parameterizedArray) {
@@ -50,4 +63,12 @@ public abstract class ShardingSphereAdapterContainer extends ShardingSphereConta
      */
     public abstract DataSource getDataSource();
     
+    @Getter
+    @RequiredArgsConstructor
+    public static class Authentication {
+        
+        private final String user;
+        
+        private final String password;
+    }
 }
