+++ shardingsphere-test/shardingsphere-integration-test/shardingsphere-integration-test-fixture/src/test/java/org/apache/shardingsphere/test/integration/junit/container/storage/impl/H2Container.java
@@ -17,44 +17,29 @@
 
 package org.apache.shardingsphere.test.integration.junit.container.storage.impl;
 
-import lombok.SneakyThrows;
 import org.apache.shardingsphere.infra.database.type.dialect.H2DatabaseType;
-import org.apache.shardingsphere.test.integration.env.datasource.builder.ActualDataSourceBuilder;
 import org.apache.shardingsphere.test.integration.junit.container.storage.ShardingSphereStorageContainer;
 import org.apache.shardingsphere.test.integration.junit.param.model.ParameterizedArray;
 
-import javax.sql.DataSource;
-import java.util.Map;
+import java.util.Objects;
 
 /**
  * H2 container.
  */
 public final class H2Container extends ShardingSphereStorageContainer {
     
-    private volatile Map<String, DataSource> actualDataSources;
-    
-    private volatile boolean isHealthy;
-    
     public H2Container(final ParameterizedArray parameterizedArray) {
         super("h2-embedded", "h2:fake", new H2DatabaseType(), true, parameterizedArray);
     }
     
     @Override
-    @SneakyThrows
-    protected void configure() {
-        super.configure();
-        actualDataSources = ActualDataSourceBuilder.createActualDataSources(getParameterizedArray().getScenario(), getParameterizedArray().getDatabaseType());
-        isHealthy = true;
-    }
-    
-    @Override
     public boolean isHealthy() {
-        return isHealthy;
+        return true;
     }
     
     @Override
     protected String getUrl(final String dataSourceName) {
-        return null;
+        return String.format("jdbc:h2:mem:%s;DB_CLOSE_DELAY=-1;DATABASE_TO_UPPER=false;MODE=MySQL", Objects.isNull(dataSourceName) ? "test_db" : dataSourceName);
     }
     
     @Override
@@ -64,17 +49,12 @@ public final class H2Container extends ShardingSphereStorageContainer {
     
     @Override
     protected String getUsername() {
-        return null;
+        return "sa";
     }
     
     @Override
     protected String getPassword() {
-        return null;
+        return "";
     }
     
-    @Override
-    @SneakyThrows
-    public synchronized Map<String, DataSource> getDataSourceMap() {
-        return actualDataSources;
-    }
 }
